+++ src/display/context.h
@@ -28,34 +28,44 @@
 #define ARK_DISPLAY_CONTEXT_H
 
 #include <stddef.h>
-#include <stdint.h>
+
+#include "platform.h"
 
 ////////////////////////////////////////////////////////////////////////////////
-static const size_t DISPLAY_CTX_STEP_COUNT          = 5;
-static const size_t DISPLAY_CTX_OP_SIZE             = 18;
-static const size_t DISPLAY_CTX_TITLE_SIZE          = DISPLAY_CTX_OP_SIZE - 3;
-static const size_t DISPLAY_CTX_TEXT_SIZE           = 68;
-static const size_t DISPLAY_CTX_EXTENDED_TEXT_SIZE  = 256;
+static const size_t DISPLAY_CTX_OP_SIZE         = 18;
+static const size_t DISPLAY_CTX_TEXT_SIZE_EXT   = 256;
+
+#if defined(SUPPORTS_LARGE_OPERATIONS)
+    static const size_t DISPLAY_CTX_STEP_COUNT  = 35;
+    static const size_t DISPLAY_CTX_TITLE_SIZE  = DISPLAY_CTX_OP_SIZE - 1;
+    static const size_t DISPLAY_CTX_TEXT_SIZE   = 130;
+#else
+    static const size_t DISPLAY_CTX_STEP_COUNT  = 5;
+    static const size_t DISPLAY_CTX_TITLE_SIZE  = DISPLAY_CTX_OP_SIZE - 3;
+    static const size_t DISPLAY_CTX_TEXT_SIZE   = 68;
+#endif
 
 ////////////////////////////////////////////////////////////////////////////////
-// Display Context.
+// Display Storage
 //
-// 704 Bytes (64-byte-aligned)
+// NanoS - 704-byte-aligned
+// NanoX - 2496-byte-aligned
+//
+// ---
 typedef struct display_context_t {
-    uint8_t     operation   [DISPLAY_CTX_OP_SIZE];
-    uint8_t     title       [DISPLAY_CTX_STEP_COUNT] [DISPLAY_CTX_TITLE_SIZE];
-    uint8_t     text        [DISPLAY_CTX_STEP_COUNT] [DISPLAY_CTX_TEXT_SIZE];
+    char    operation   [DISPLAY_CTX_OP_SIZE];
+    char    title       [DISPLAY_CTX_STEP_COUNT] [DISPLAY_CTX_TITLE_SIZE];
+    char    text        [DISPLAY_CTX_STEP_COUNT] [DISPLAY_CTX_TEXT_SIZE];
 
-    // For potentially large display texts.
-    // Should always be set as last step in the UX display flow.
+    // For large display texts.
+    // Should always be the last step in a UX display flow.
     //
-    // 255 strLen + the null-terminator.
     // e.g. Message, VendorField, IPFS DAG
-    uint8_t extended_title[DISPLAY_CTX_TITLE_SIZE];
-    uint8_t extended_text[DISPLAY_CTX_EXTENDED_TEXT_SIZE];
+    char    title_ext   [DISPLAY_CTX_TITLE_SIZE];
+    char    text_ext    [DISPLAY_CTX_TEXT_SIZE_EXT];
 } DisplayContext;
 
 ////////////////////////////////////////////////////////////////////////////////
 extern DisplayContext displayCtx;
 
-#endif  // #define ARK_DISPLAY_CONTEXT_H
+#endif  // ARK_DISPLAY_CONTEXT_H