+++ src/operations/transactions/ux/display_ux.c
@@ -27,6 +27,7 @@
 #include "transactions/ux/display_ux.h"
 
 #include <stdbool.h>
+#include <stddef.h>
 #include <string.h>
 
 #include "transactions/transaction.h"
@@ -45,51 +46,48 @@
 #include "display/display.h"
 
 ////////////////////////////////////////////////////////////////////////////////
-extern void setDisplaySteps(uint8_t steps, bool isExtended);
+extern void SetUxDisplay(size_t steps, bool isExtended);
 
 ////////////////////////////////////////////////////////////////////////////////
-void setDisplay(const Transaction *transaction) {
+void SetUx(const Transaction *transaction) {
     explicit_bzero(&displayCtx, sizeof(displayCtx));
 
     const bool hasVendorField = transaction->vendorFieldLength > 0;
 
     switch (transaction->type) {
         case TRANSFER_TYPE:
-            displayTransfer(transaction);
-            
-            setDisplaySteps(STEPS_TRANSFER_MIN + (uint8_t)hasVendorField,
-                            hasVendorField);
+            SetUxTransfer(transaction);
+            SetUxDisplay(UX_TRANSFER_STEPS + hasVendorField,hasVendorField);
             break;
 
         case SECOND_SIGNATURE_TYPE:
-            displaySecondSignature(transaction);
-            setDisplaySteps(STEPS_SECOND_SIGNATURE, false);
+            SetUxSecondSignature(transaction);
+            SetUxDisplay(UX_SECOND_SIGNATURE_STEPS, false);
             break;
 
         case VOTE_TYPE:
-            displayVote(transaction);
-            setDisplaySteps(STEPS_VOTE, false);
+            SetUxVote(transaction);
+            SetUxDisplay(UX_VOTE_STEPS, false);
             break;
 
         case IPFS_TYPE:
-            displayIpfs(transaction);
-            setDisplaySteps(STEPS_IPFS, true);
+            SetUxIpfs(transaction);
+            SetUxDisplay(UX_IPFS_STEPS, true);
             break;
 
         case HTLC_LOCK_TYPE:
-            displayHtlcLock(transaction);
-            setDisplaySteps(STEPS_HTLC_LOCK_MIN + (uint8_t)hasVendorField,
-                            hasVendorField);
+            SetUxHtlcLock(transaction);
+            SetUxDisplay(UX_HTLC_LOCK_STEPS + hasVendorField, hasVendorField);
             break;
 
         case HTLC_CLAIM_TYPE:
-            displayHtlcClaim(transaction);
-            setDisplaySteps(STEPS_HTLC_CLAIM, false);
+            SetUxHtlcClaim(transaction);
+            SetUxDisplay(UX_HTLC_CLAIM_STEPS, false);
             break;
 
         case HTLC_REFUND_TYPE:
-            displayHtlcRefund(transaction);
-            setDisplaySteps(STEPS_HTLC_REFUND, false);
+            SetUxHtlcRefund(transaction);
+            SetUxDisplay(UX_HTLC_REFUND_STEPS, false);
             break;
 
         default: break;