+++ src/utils/print.h
@@ -31,11 +31,11 @@
 #include <stdint.h>
 
 ////////////////////////////////////////////////////////////////////////////////
-size_t printAmount(uint64_t amount,
-                   uint8_t *out,
-                   size_t outSize,
-                   const char *tokenName,
-                   size_t tokenNameSize,
-                   uint8_t decimals);
+size_t UintToString(uint64_t value, char *dst, size_t maxLen);
 
-#endif  // #ifndef ARK_UTILS_PRINT_H
+////////////////////////////////////////////////////////////////////////////////
+size_t TokenAmountToString(const char *token, size_t tokenLen, size_t decimals,
+                           uint64_t amount,
+                           char *dst, size_t maxLen);
+
+#endif  // ARK_UTILS_PRINT_H